%.* require a int not unsigned

+++ server/dhcp.c
@@ -1053,9 +1053,9 @@ void dhcpinform (packet, ms_nulltp)
 		i = d1.len;
 		if (i >= sizeof(raw.file)) {
 			log_info("file name longer than packet field "
-				 "truncated - field: %lu name: %d %.*s", 
-				 (unsigned long)sizeof(raw.file), i, i,
-				 d1.data);
+				 "truncated - field: %lu name: %u %.*s", 
+				 (unsigned long)sizeof(raw.file), i,
+				 (int)i, d1.data);
 			i = sizeof(raw.file);
 		} else
 			raw.file[i] = 0;
@@ -2877,10 +2877,10 @@ void dhcp_reply (lease)
 				(sizeof raw.sname) - state -> server_name.len);
 		else 
 			log_info("server name longer than packet field "
-				 "truncated - field: %lu name: %d %.*s", 
+				 "truncated - field: %lu name: %u %.*s", 
 				 (unsigned long)sizeof(raw.sname),
 				 state->server_name.len,
-				 state->server_name.len,
+				 (int)state->server_name.len,
 				 state->server_name.data);
 	} else
 		bufs |= 2; /* XXX */
